+++ src/Core/Dfe.Complete.Application/LocalAuthorities/Queries/GetLocalAuthority/GetLocalAuthorityByCodeQueryHandler.cs
@@ -3,12 +3,13 @@ using Dfe.Complete.Domain.Entities;
 using Dfe.Complete.Domain.Interfaces.Repositories;
 using Dfe.Complete.Application.LocalAuthorities.Models;
 using MediatR;
+using Microsoft.Extensions.Logging;
 
 namespace Dfe.Complete.Application.LocalAuthorities.Queries.GetLocalAuthority;
 
 public record GetLocalAuthorityByCodeQuery(string Code) : IRequest<Result<LocalAuthorityDto?>>;
 
-public class GetLocalAuthorityByCodeQueryHandler(ICompleteRepository<LocalAuthority> localAuthorityRepo)
+public class GetLocalAuthorityByCodeQueryHandler(ICompleteRepository<LocalAuthority> localAuthorityRepo, ILogger<GetLocalAuthorityByCodeQueryHandler> logger)
     : IRequestHandler<GetLocalAuthorityByCodeQuery, Result<LocalAuthorityDto?>>
 {
     public async Task<Result<LocalAuthorityDto?>> Handle(GetLocalAuthorityByCodeQuery request, CancellationToken cancellationToken)
@@ -23,6 +24,7 @@ public class GetLocalAuthorityByCodeQueryHandler(ICompleteRepository<LocalAuthor
         }
         catch (Exception e)
         {
+            logger.LogError(e, "Exception for {Name} Request - {@Request}", nameof(GetLocalAuthorityByCodeQueryHandler), request);
             return Result<LocalAuthorityDto?>.Failure(e.Message);
         }
     }