+++ src/Core/Dfe.Complete.Application/Projects/Queries/GetProject/GetProjectGroupByGroupReferenceNumber.cs
@@ -4,13 +4,15 @@ using Dfe.Complete.Application.Common.Models;
 using Dfe.Complete.Application.Projects.Models;
 using Dfe.Complete.Domain.Entities;
 using AutoMapper;
+using Microsoft.Extensions.Logging;
 
 namespace Dfe.Complete.Application.Projects.Queries.GetProject
 {
     public record GetProjectGroupByGroupReferenceNumberQuery(string GroupReferenceNumber) : IRequest<Result<ProjectGroupDto>>;
 
     public class GetProjectGroupByGroupReferenceNumberQueryHandler(ICompleteRepository<ProjectGroup> projectGroupRepository,
-        IMapper mapper)
+        IMapper mapper,
+        ILogger<GetProjectGroupByGroupReferenceNumberQueryHandler> logger)
         : IRequestHandler<GetProjectGroupByGroupReferenceNumberQuery, Result<ProjectGroupDto>>
     {
         public async Task<Result<ProjectGroupDto?>> Handle(GetProjectGroupByGroupReferenceNumberQuery request, CancellationToken cancellationToken)
@@ -25,6 +27,7 @@ namespace Dfe.Complete.Application.Projects.Queries.GetProject
             }
             catch (Exception ex)
             {
+                logger.LogError(ex, "Exception for {Name} Request - {@Request}", nameof(GetProjectGroupByGroupReferenceNumberQueryHandler), request);
                 return Result<ProjectGroupDto?>.Failure(ex.Message);
             }