+++ src/Frontend/Dfe.Complete/Pages/Projects/CreateNewProject.cshtml.cs
@@ -9,7 +9,7 @@ namespace Dfe.Complete.Pages.Projects
         public string? ProjectType { get; set; }
         
         public IActionResult OnPost()
-        {
+        { 
             var pageToRedirectTo = ProjectType switch
             {
                 "conversion" => "/Projects/Conversion/CreateNewProject",